Extend Rspec service from Runner and move generic logic there
