+++ lib/upsteem/deploy/services/test_suite_runners.rb
@@ -1,6 +1,7 @@
 %w[
   base
   skipper
+  runner
   rspec
 ].each do |file|
   require_relative("test_suite_runners/#{file}")
